requirements.txt
fastapi
uvicorn
playwright
bertopic
kiwipiepy
openai
python-dotenv
pandas
numpy
scikit-learn
pymongo
